--- PlanetDB.xcodeproj/xcuserdata/Antonio.xcuserdatad/xcschemes/xcschememanagement.plist
@@ -1,27 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<!DOCTYPE plist PUBLIC "-//Apple//DTD PLIST 1.0//EN" "http://www.apple.com/DTDs/PropertyList-1.0.dtd">
-<plist version="1.0">
-<dict>
-	<key>SchemeUserState</key>
-	<dict>
-		<key>PlanetDB.xcscheme</key>
-		<dict>
-			<key>orderHint</key>
-			<integer>1</integer>
-		</dict>
-	</dict>
-	<key>SuppressBuildableAutocreation</key>
-	<dict>
-		<key>A9FDB7311E44044F00D08A38</key>
-		<dict>
-			<key>primary</key>
-			<true/>
-		</dict>
-		<key>A9FDB7451E44044F00D08A38</key>
-		<dict>
-			<key>primary</key>
-			<true/>
-		</dict>
-	</dict>
-</dict>
-</plist>
