--- PlanetDB/Assets.xcassets/Sample Images/Contents.json
@@ -1,6 +1,0 @@
-{
-  "info" : {
-    "version" : 1,
-    "author" : "xcode"
-  }
-}
