--- PlanetDB/MealTableViewCell.swift
@@ -1,28 +1,0 @@
-//
-//  MealTableViewCell.swift
-//  PlanetDB
-//
-//  Created by Antonio Medrano on 4/26/17.
-//  Copyright © 2017 Antonio Medrano. All rights reserved.
-//
-
-import UIKit
-
-class MealTableViewCell: UITableViewCell {
-
-    //MARK: Properties
-    @IBOutlet weak var nameLabel: UILabel!
-    @IBOutlet weak var photoImageView: UIImageView!
-    @IBOutlet weak var ratingControl: RatingControl!
-    
-    override func awakeFromNib() {
-        super.awakeFromNib()
-        // Initialization code
-    }
-
-    override func setSelected(_ selected: Bool, animated: Bool) {
-        super.setSelected(selected, animated: animated)
-
-        // Configure the view for the selected state
-    }
-}
