+++ PlanetDB/PlanetDB/ViewController.swift
@@ -1,0 +1,25 @@
+//
+//  ViewController.swift
+//  PlanetDB
+//
+//  Created by Antonio Medrano on 4/27/17.
+//  Copyright © 2017 Antonio Medrano. All rights reserved.
+//
+
+import UIKit
+
+class ViewController: UIViewController {
+
+    override func viewDidLoad() {
+        super.viewDidLoad()
+        // Do any additional setup after loading the view, typically from a nib.
+    }
+
+    override func didReceiveMemoryWarning() {
+        super.didReceiveMemoryWarning()
+        // Dispose of any resources that can be recreated.
+    }
+
+
+}
+
